requirements.txt
tqdm
spconv==2.3.6
torchsparse @ git+https://github.com/mit-han-lab/torchsparse.git@v1.4.0
MinkowskiEngine==0.5.4
torch-scatter==2.1.1
torch-geometric==2.3.1
nuscenes-devkit
easydict
hydra-core==1.3.2
pytorch-lightning==1.6.5